+++ lib/apathy_drive/systems/death.ex
@@ -16,12 +16,16 @@ defmodule Systems.Death do
 
     send(spirit.socket_pid, {:set_entity, spirit})
 
+    MonsterTemplate.set_last_killed_at(monster)
+
     ApathyDrive.Repo.delete(monster)
     Process.exit(self, :normal)
   end
 
   def kill(%Monster{spirit: nil} = monster) do
     ApathyDrive.PubSub.broadcast!("rooms:#{monster.room_id}", {:monster_died, monster: monster, reward: monster.experience})
+
+    MonsterTemplate.set_last_killed_at(monster)
 
     ApathyDrive.Repo.delete(monster)
     Process.exit(self, :normal)