+++ priv/repo/migrations/20150502203405_add_last_killed_at_to_monster_templates.exs
@@ -1,0 +1,9 @@
+defmodule ApathyDrive.Repo.Migrations.AddLastKilledAtToMonsterTemplates do
+  use Ecto.Migration
+
+  def change do
+    alter table(:monster_templates) do
+      add :last_killed_at, :datetime
+    end
+  end
+end