+++ priv/repo/migrations/20150502211202_add_regen_time_to_monster_templates.exs
@@ -1,0 +1,9 @@
+defmodule ApathyDrive.Repo.Migrations.AddRegenTimeToMonsterTemplates do
+  use Ecto.Migration
+
+  def change do
+    alter table(:monster_templates) do
+      add :regen_time_in_minutes, :integer
+    end
+  end
+end